# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API Server for Suasor",
        "title": "Suasor API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Dane Miller",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/clients/media/{clientID}/series/{seriesID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a specific TV series from the client by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "series"
                ],
                "summary": "Get series by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Client ID",
                        "name": "clientID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Series ID",
                        "name": "seriesID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Series retrieved",
                        "schema": {
                            "$ref": "#/definitions/responses.APIResponse-responses_MediaItemResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid client ID",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    }
                }
            }
        },
        "/clients/media/{clientID}/series/{seriesID}/seasons": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves all seasons for a specific TV series",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "series"
                ],
                "summary": "Get seasons for a series",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Client ID",
                        "name": "clientID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Series ID",
                        "name": "seriesID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Seasons retrieved",
                        "schema": {
                            "$ref": "#/definitions/responses.APIResponse-array_responses_MediaItemResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid client ID",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    }
                }
            }
        },
...
        "/series/actor/{actor}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves TV series featuring a specific actor",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "series"
                ],
                "summary": "Get series by actor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Actor name",
                        "name": "actor",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Series retrieved",
                        "schema": {
                            "$ref": "#/definitions/responses.APIResponse-array_responses_MediaItemResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    }
                }
            }
        },
        "/series/creator/{creator}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves TV series by a specific creator/director",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "series"
                ],
                "summary": "Get series by creator",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Creator name",
                        "name": "creator",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Series retrieved",
                        "schema": {
                            "$ref": "#/definitions/responses.APIResponse-array_responses_MediaItemResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    }
                }
            }
        },
        "/series/genre/{genre}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves TV series from all connected clients that match the specified genre",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "series"
                ],
                "summary": "Get series by genre",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Genre name",
                        "name": "genre",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Series retrieved",
                        "schema": {
                            "$ref": "#/definitions/responses.APIResponse-array_responses_MediaItemResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse-responses_ErrorDetails"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "responses.APIResponse-array_responses_MediaItemResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.MediaItemResponse"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Operation successful"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "responses.APIResponse-responses_MediaItemResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/responses.MediaItemResponse"
                },
                "message": {
                    "type": "string",
                    "example": "Operation successful"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "responses.ErrorResponse-responses_ErrorDetails": {
            "type": "object",
            "properties": {
                "details": {
                    "$ref": "#/definitions/responses.ErrorDetails"
                },
                "message": {
                    "type": "string",
                    "example": "This is a pretty message"
                },
                "request_id": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer",
                    "example": 201
                },
                "timestamp": {
                    "type": "string"
                },
                "type": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/errors.ErrorType"
                        }
                    ],
                    "example": "FAILED_CHECK"
                }
            }
        },
...
        "responses.MediaItemResponse": {
            "type": "object",
            "properties": {
                "clientId": {
                    "type": "integer"
                },
                "clientType": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "data": {},
                "externalId": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "type": {
                    "$ref": "#/definitions/types.MediaType"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    }
}